Fix type comparison

+++ src/displayapp/screens/settings/SettingShortcutApp.cpp
@@ -56,7 +56,7 @@ std::unique_ptr<Screen> SettingShortcutApp::CreateScreen(unsigned int screenNum)
 
   auto selectedApp = settingsController.GetShortcutApp();
   uint32_t selectedIndex = 0;
-  for (int i = 0; i < UserAppTypes::Count; i++) {
+  for (size_t i = 0; i < UserAppTypes::Count; i++) {
     if (Pinetime::Applications::userApps[i].app == selectedApp) {
       selectedIndex = i + 1;
       break;
